libhsakmt: modify XNACK mode API
XNACK mode modified to not return a list of GPUs but
instead, fail to set XNACK mode chage if any GFXv9+
visible to the process can not support it.

This will allow to always enable SVM on all GFXv9 or later
but may fail to switch to the requested XNACK mode instead.

Change-Id: Ifcd56720c76a4d0e7e4b39dd4d0af3643f44418e
Signed-off-by: Alex Sierra <alex.sierra@amd.com>